requirements.txt
pillow
google-generativeai==0.5.2
IPython